fixing trivial warnings

+++ src/common/SharedMemoryCache.cpp
@@ -158,7 +158,7 @@ SharedMemoryCache::~SharedMemoryCache() {
     if (false) {
         //code from FileCacheRegister...
     }
-    for (int i = 0; i < _numBlocks; i++) {
+    for (size_t i = 0; i < _numBlocks; i++) {
         if (_blkIndex[i].activeCnt > 0) {
             std::cout << "[TAZER] " << _name << " " << i << " " << _numBlocks << " " << _blkIndex[i].activeCnt << " " << _blkIndex[i].fileIndex - 1 << " " << _blkIndex[i].blockIndex - 1 << " "
                       << "prefetched" << _blkIndex[i].prefetched << std::endl;
@@ -237,7 +237,7 @@ std::vector<std::shared_ptr<BoundedCache<MultiReaderWriterLock>::BlockEntry>> Sh
     // memcpy(entries, &_blkIndex[binIndex * _associativity], sizeof(BlockEntry) * _associativity);
     std::vector<std::shared_ptr<BlockEntry>> entries;
     int startIndex = binIndex * _associativity;
-    for (int i = 0; i < _associativity; i++) {
+    for (size_t i = 0; i < _associativity; i++) {
         // entries.emplace_back(std::make_shared<BlockEntry>(*(BlockEntry*)&_blkIndex[i+startIndex]));
         entries.emplace_back(new MemBlockEntry());
         *entries[i].get() = *(BlockEntry *)&_blkIndex[i + startIndex];
